Return the password in the delete_account api

+++ app.py
@@ -768,14 +768,14 @@ def get_all_contacts(current_user):
 @token_required
 def delete_account(current_user):
     data = request.get_json()
-    #password = data.get('password')
-    #checks if password exists in data
-    # if not password:
-    #     return jsonify({'error': 'Password is required to delete the account.'}), 400
+    password = data.get('password')
+    checks if password exists in data
+    if not password:
+        return jsonify({'error': 'Password is required to delete the account.'}), 400
 
-    #checks with the password in the database
-    # if not check_password_hash(current_user['password'], password):
-    #     return jsonify({'error': 'Incorrect password.'}), 400
+    checks with the password in the database
+    if not check_password_hash(current_user['password'], password):
+        return jsonify({'error': 'Incorrect password.'}), 400
 
     journal_collection.delete_many({'email': current_user['email']})
     users_collection.delete_one({'username': current_user['username']})
